+++ frontend/src/employee-frontend/components/fee-decisions/FeeDecisionsPage.tsx
@@ -5,7 +5,7 @@
 import React, { useCallback, useContext, useEffect, useState } from 'react'
 import { Gap } from 'lib-components/white-space'
 import { Container, ContentArea } from 'lib-components/layout/Container'
-import { Checked, InvoicingUiContext } from '../../state/invoicing-ui'
+import { InvoicingUiContext } from '../../state/invoicing-ui'
 import FeeDecisions from './FeeDecisions'
 import FeeDecisionFilters from './FeeDecisionFilters'
 import Actions from './Actions'
@@ -19,6 +19,7 @@ import {
 import { useRestApi } from 'lib-common/utils/useRestApi'
 import { FeeDecisionSummary } from '../../types/invoicing'
 import { SearchOrder } from '../../types'
+import { useCheckedState } from '../../state/invoicing'
 
 const pageSize = 200
 
@@ -52,20 +53,7 @@ const FeeDecisionsPage = React.memo(function FeeDecisionsPage() {
     feeDecisions: { searchFilters, debouncedSearchTerms }
   } = useContext(InvoicingUiContext)
 
-  const [checked, setChecked] = useState<Checked>({})
-  const toggleChecked = (id: string) =>
-    setChecked({
-      ...checked,
-      [id]: !checked[id]
-    })
-  const checkIds = (ids: string[]) => {
-    const idsChecked = ids.map((id) => ({ [id]: true }))
-    setChecked({
-      ...checked,
-      ...Object.assign({}, ...idsChecked)
-    })
-  }
-  const clearChecked = () => setChecked({})
+  const checkedState = useCheckedState()
 
   const loadDecisions = useCallback(() => {
     const status = searchFilters.status
@@ -105,11 +93,13 @@ const FeeDecisionsPage = React.memo(function FeeDecisionsPage() {
   const checkAll = useCallback(() => {
     const currentPage = decisions[page]
     if (currentPage.isSuccess) {
-      checkIds(currentPage.value.map((decision) => decision.id))
+      checkedState.checkIds(currentPage.value.map((decision) => decision.id))
     }
-  }, [decisions, checkIds]) // eslint-disable-line react-hooks/exhaustive-deps
+  }, [decisions, checkedState.checkIds]) // eslint-disable-line react-hooks/exhaustive-deps
 
-  const checkedIds = Object.keys(checked).filter((id) => !!checked[id])
+  const checkedIds = Object.keys(checkedState.checked).filter(
+    (id) => !!checkedState.checked[id]
+  )
 
   return (
     <Container data-qa="fee-decisions-page">
@@ -130,16 +120,16 @@ const FeeDecisionsPage = React.memo(function FeeDecisionsPage() {
           sortDirection={sortDirection}
           setSortDirection={setSortDirection}
           showCheckboxes={searchFilters.status === 'DRAFT'}
-          checked={checked}
-          toggleChecked={toggleChecked}
+          checked={checkedState.checked}
+          toggleChecked={checkedState.toggleChecked}
           checkAll={checkAll}
-          clearChecked={clearChecked}
+          clearChecked={checkedState.clearChecked}
         />
       </ContentArea>
       <Actions
         status={searchFilters.status}
         checkedIds={checkedIds}
-        clearChecked={clearChecked}
+        clearChecked={checkedState.clearChecked}
         loadDecisions={loadDecisions}
       />
     </Container>
